Add code that enables the verification of strings without the need for exact coordinates. Instead, the buffer is captured and put into a list. This list can be iterated over to find the string you're looking for.
The reason behind implementing such a basic and admittedly cludgy way of finding
a string in the 3270 terminal buffer is that at IBM not all of our test software
uses interfaces where exact locations of text elements would be known. Therefore
it is easier to just capture the screen, and try to find the expected output in
the captured buffer.